# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/users":{"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["users"]},"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["users"]}},"/users/{id}":{"put":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["users"]},"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["users"]}},"/users/upload/{id}":{"post":{"operationId":"UsersController_updateWithImage","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["users"]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterAuthDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginAuthDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/roles":{"post":{"operationId":"RolesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRolDto"}}}},"responses":{"201":{"description":""}},"tags":["roles"]}},"/categories":{"get":{"operationId":"CategoriesController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["categories"]},"post":{"operationId":"CategoriesController_createWithImage","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCategoryDto"}}}},"responses":{"201":{"description":""}},"tags":["categories"]}},"/categories/{id}":{"put":{"operationId":"CategoriesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCategoryDto"}}}},"responses":{"200":{"description":""}},"tags":["categories"]},"delete":{"operationId":"CategoriesController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["categories"]}},"/categories/upload/{id}":{"put":{"operationId":"CategoriesController_updateWithImage","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCategoryDto"}}}},"responses":{"200":{"description":""}},"tags":["categories"]}},"/products":{"get":{"operationId":"ProductsController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["products"]},"post":{"operationId":"ProductsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductDto"}}}},"responses":{"201":{"description":""}},"tags":["products"]}},"/products/{id}":{"get":{"operationId":"ProductsController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["products"]},"put":{"operationId":"ProductsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductDto"}}}},"responses":{"200":{"description":""}},"tags":["products"]},"delete":{"operationId":"ProductsController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["products"]}},"/products/category/{id_category}":{"get":{"operationId":"ProductsController_findByCategory","parameters":[{"name":"id_category","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["products"]}},"/products/search/{name}":{"get":{"operationId":"ProductsController_findByName","parameters":[{"name":"name","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["products"]}},"/products/upload/{id}":{"put":{"operationId":"ProductsController_updateWithImage","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductDto"}}}},"responses":{"200":{"description":""}},"tags":["products"]}},"/address":{"post":{"operationId":"AddressController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAddressDto"}}}},"responses":{"201":{"description":""}},"tags":["address"]},"get":{"operationId":"AddressController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["address"]}},"/address/user/{id_user}":{"get":{"operationId":"AddressController_findByUser","parameters":[{"name":"id_user","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["address"]}},"/address/{id}":{"put":{"operationId":"AddressController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAddressDto"}}}},"responses":{"200":{"description":""}},"tags":["address"]},"delete":{"operationId":"AddressController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["address"]}},"/mercadopago/identification_types":{"get":{"operationId":"MercadoPagoController_getIdentificationTypes","parameters":[],"responses":{"200":{"description":""}},"tags":["mercadopago"]}},"/mercadopago/installments/{first_six_digits}/{amount}":{"get":{"operationId":"MercadoPagoController_getInstallments","parameters":[{"name":"first_six_digits","required":true,"in":"path","schema":{"type":"number"}},{"name":"amount","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["mercadopago"]}},"/mercadopago/card_token":{"post":{"operationId":"MercadoPagoController_createCardToken","parameters":[],"responses":{"201":{"description":""}},"tags":["mercadopago"]}},"/mercadopago/payments":{"post":{"operationId":"MercadoPagoController_createPayment","parameters":[],"responses":{"201":{"description":""}},"tags":["mercadopago"]}},"/orders":{"get":{"operationId":"OrdersController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["orders"]}},"/orders/{id_client}":{"get":{"operationId":"OrdersController_findByClient","parameters":[{"name":"id_client","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["orders"]}},"/orders/update-dispatched/{id}":{"put":{"operationId":"OrdersController_updateStatus","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["orders"]}}},"info":{"title":"Ansur backend","description":"Descripcion de las APIs","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{}},"UpdateUserDto":{"type":"object","properties":{}},"RegisterAuthDto":{"type":"object","properties":{}},"LoginAuthDto":{"type":"object","properties":{}},"CreateRolDto":{"type":"object","properties":{}},"CreateCategoryDto":{"type":"object","properties":{}},"UpdateCategoryDto":{"type":"object","properties":{}},"CreateProductDto":{"type":"object","properties":{}},"UpdateProductDto":{"type":"object","properties":{}},"CreateAddressDto":{"type":"object","properties":{}},"UpdateAddressDto":{"type":"object","properties":{}}}}}